Use the boeotian helmet consistently everywhere. This revealed some huge duplication in files,
which cannot be fixed without hacks such as models with only props and a
small triangle
Move all the modified helmets to their new folder, so it's easier to
look for them in atlas
Add a new tin texture with and without laurels to fit the new uvmap Move
the boeotian helmet
 from  the head prop point to the helmet prop point based on a patch
from WowGetOffYourCellPhone
Make the boeotian helmet more modular using props so variants are easier
to make for modders,
nuke the remaining models.

This was SVN commit r20702.