Gather near dropsite rather than current position after finishing construction
Summary:
{F316877}

Often we'll have a large group of units that we want to move to a new
location, but if we do that by having them building a dropsite, then if
the building's finished before everyone gets there, units still en route
will start gathering near their current position instead of near the new
location.

Test Plan: Agree.

Reviewers: wraitii

Reviewed By: wraitii
Subscribers: wraitii, Vulcan

Differential Revision: https://code.wildfiregames.com/D1162
This was SVN commit r20876.