Fix AuraManager OOS on rejoin following c90d72deb5 / D1108.
The AuraManager serialized the affects array of the aura JSON files,
while the Auras components create new affects arrays upon
deserialization.
So the AuraManager of rejoined clients failed to identify and remove the
according aura effects of units who received the aura effect before the
rejoin.

Fix it by not serializing the array but only the strings inside the
array.
It were preferable to rebuild the modification cache and remove the
serialization altogether, refs #3834.

f24523dc8f introduced the same OOS on rejoin, refs #3909, #4239, but it
wasn't noticed due to the revert in d7e51f2a4f.

Differential Revision: https://code.wildfiregames.com/D1201
Reviewed By: bb
Comments By: gameboy
Fixes #4924

This was SVN commit r21014.