Cleanup random map Area prototype following b4503bb61e.
Delete areaID and just identify the Area object by the actual JS object
reference.
Delete all references from Areas in the RandomMap object,
especially the area property of the RandomMap object became invalid
after creating a new area.
This was replaced with a cache in the Area object and a contains
function.
Add getPoints getter, so that the Area property is not accessed
directly.
Deepfreeze points to prevent accidental overwrites.
Implement StayAreasConstraint and extend currently unused
AvoidAreaConstraint to accept multiple areas.

This was SVN commit r21296.