Unify the two copies of the custom dock placement logic on Hellas and Pompeii, refs #4805.
Extend it to only place docks where the according body of water has a
minimum size, fixes #5065.
Compute the dock angle similar to GetDockAngle from Commands.js (as the
previous angle computation is wrong if the large body of water doesn't
lie in that direction).

This was SVN commit r21433.