Rename some heroes using the common scheme.
Those heroes were introduced in 097c19a6a0
The new convention from now on will be civ_hero_name(_type(_weapontype))
where parenthesis are optionnal.
However it would be nice to when a infantry hero is added for instance,
than the other one specify it's cavalry
Since it requires map changes, it will be up to the person introducing
it.

This was SVN commit r21449.