Remove some duplication in the Jebel Barkal template loading.
Prevent crash of the map when an odd playercount was chosen due to
primeSortPlayers returning a wrong number of elements as reported by
temple.
Remove some TODOs that likely won't be done soon and don't seem crucial.
Fix missing atlas biome rename in Aethiopia / Fields of Meroe.

This was SVN commit r21531.