Don't have the difficulty of Jebel Barkal differ by the factor 10 depending on the mapsize and playercount.
If there are less spawn points, the groups of spawned units is increased
proportionally.
The new balancing should yield the same results as a 1v1 on medium
mapsize or a 4v4 on a normal mapsize in the previous balancing.

Notice that the number of elephant stables (and hence elephants) can
still vary.

This was SVN commit r21613.