Improve mine and passage placement on Jebel Barkal.
Place more mines on nomad in the fertile land.
Chose better fitting templates for the fertile land.
Alternate passage locations, so the gameplay purpose of the irrigation
canals is met (increasing the distance between opponents).
Increase the size of the city on larger maps.
Don't have bumps near gates.

This was SVN commit r21743.