Fix gaia attackers on the left riverside of Danubius trying to attack units on the right riverside if they happened to be closer, fixes #4622.
At most send an attack order for one target before patroling
(targetCount),
because adding further targets can end in the gaia unit being blocked
and slaughtered by the player without difficulty and
because the mechanism isn't as relevant for ships as it was envisioned
to be (since most ofen the 3 closest ships already are in the same
location).

Patch By: (-_-) aka smiley
Differential Revision: https://code.wildfiregames.com/D1449
This was SVN commit r21767.