Correct Mauryans into Mauryas.
The name of the faction is the dynasty name "Maurya". The plural should
be "Maurya" as well, but "Mauryas" is commonly accepted and less likely
to confuse people. On the other hand, "Mauryan" is the adjective, and
cannot get a plural mark in English.
All instances of "Mauryans" were removed, including in file names and
paths.

Patch By: Nescio
Differential Revision: https://code.wildfiregames.com/D1342
This was SVN commit r22005.