Move manual page values back to the page itself due to a lack of that page being reusable.
Instead encourage developers to create a custom GUI page for their
needs, since that is cheap and since they won't be stuck with this
inextensible page.

0da7e822ff reused the manual page to display userfeedback terms.
dfacf5358c moved more values to the calling pages and introduced a Title
string that is never displayed.
da49aa1541 stopped using the manual page to display terms, because one
could actually not reuse the page to do something more than just
displaying text and using one website link.

refs D1684

This was SVN commit r22665.