Fix most of the warnings.
The last one needs a png file see
https://wildfiregames.com/forum/index.php?/topic/26720-task-aspis-remake/page/3/&tab=comments#comment-383062

Add two incorrectly deleted animations in the new folder in 9ef5e274c7
Make the erroring variants use the animations added in 3f0dcc6353 and
cf1e1b463c instead of the 'new' folder.
Make the fisherman use the death animation.
Use the c variant of an animation instead of the d one that does not
exist.

This was SVN commit r22742.