manual/intro.txt corrections
Summary:
Another follow-up to D1719 and D1744. This one does the following:
* [font="sans-14"] (default) at the end of the line that used a
different font, rather than at the beginning of the next line (no net
change, but looks better in the file)
* two spaces for indentation of hotkeys (looks better in game)
* ` – ` (space en-dash space) instead of `: ` (colon space) to separate
hotkey and description, for better visibility and consistency.
* `6400px by 4800px` → `6400×4800 pixels`
* removed Alt+W entry (superseded by Alt+Shift+W above)
* inserted Esc entry under save game for completeness
* rephrased several descriptions

Before: {F1014718}
After: {F1014719}

Test Plan: Launch the game and open the manual.

Reviewers: O7 Catchall, Gallaecio, elexis, Itms

Reviewed By: Gallaecio
Subscribers: Freagarach, O8 JS GUI, Stan, elexis

Differential Revision: https://code.wildfiregames.com/D2110
This was SVN commit r22904.