Rewrite Savegame page to use class syntax, enabled by bdc6a7d3fd / D2290, refs #5387.
Uses simple event subscription to make the classes independent of each
other.
Removes all globals except two variables and one function (that may be
avoided in the future too).
Allows the reader to focus on the relevant aspect of the implementation
(encapsulation, information hiding) and makes implementation more
modifiable and extensible.

Differential Revision: https://code.wildfiregames.com/D2310
This was SVN commit r22985.