Fix Atlas complaining again, this time for the objectives dialog from e3f43f6352, refs 966b1608da.
The atlas check added in e3f43f6352 works for skirmish maps but not for
random maps generated in Atlas.
The same atlas check added in f2778ea7b4 works for random maps since the
partial information Atlas provides happened to be sufficient for
DiplomacyPlayerText.

Reported by Stan in 6814a983e1.

This was SVN commit r23097.