React to target death properly when chasing
Differential Revision: https://code.wildfiregames.com/D2517
Reviewed by: Freagarach
Comments by: elexis, Stan
Fixes: #5610 #5656

Introduce likelyFailure check in the movementupdate for chasing state
missing since 32e8ed51aa.
Use new order and substate for Chasing to get into the wanted range and
avoid that way issue with running - walking tracking or animation
variants.

Fixing issue when attacker would not react to the death of the target
while chasing it in stances where results for abandoning chasing are
always false and also do not rely on timer ending chasing for the rest
of stances.

This was SVN commit r23566.