Adjust the UV mapping of the "elephant_african_forest.dae" Mesh
Due to recent Elephant update for having the blend file of the whole
elephants and have the proper workspace for animating riders it was
updated in a recent commit. However the african forest elephant mesh
wasn't properly adapting to the new UV mapping since it wasn't properly
alligned.

Thread:
https://wildfiregames.com/forum/index.php?/topic/27906-elephant-uv-mapping/&tab=comments#comment-393918

This was SVN commit r23574.