Actually do not use sorting order for forced columnar formation
f19a8d6b41 moved sorting classes to template and according to lines 561
and 574 in that commit, it was intended to not use sorting in case of
forced columnar formation.

Fixing oversight at lines 660 and 662 in that commit by removing this.
word so cavalry will be put to the front always.

Noticed in D2707 after @elexis mentioned behaviour change on unused
variable.

Differential Revision: https://code.wildfiregames.com/D2711
This was SVN commit r23613.