Tweak wonder/pyramid cost and remove healing aura from wonders.
Wonders cost food previously, which is not in line with their look or
other buildings.
Their healing aura is removed to make the slightly easier to capture.

Kushites pyramids are adjusted slightly, and the monumental architecture
tech moved to the CC to match the Persian.

Patch by: Nescio
Accepted By: borg-, Feldfeld
Differential Revision: https://code.wildfiregames.com/D2686
This was SVN commit r23824.