Fix GetTechModifiedProperty mistake in 70c71bff0f & allow multiple token modifiers.
String-values in modifications fail if the modifications do not apply,
as they end up in the "multiply/add" path, following D270 / 70c71bff0f.

Also allow multiple token modifiers, so that multiple tokens may be
added for example. This becomes order-dependent but the positive
outweighs the negative.

Reported By: wowgetoffyourcellphone
Reviewed By: Angen
Differential Revision: https://code.wildfiregames.com/D2894
This was SVN commit r23862.