Disable Persian Hall & Ishtar gate as buildable structures.
The Kardakes mercenaries, trainable at the Persian Hall, are
historically problematic. Few sources exist on Kardakes troops, but they
seem to have been rather heavy infantry, and likely not mercenaries. As
such, they would be rather redundant with the Immortal.
The elephants are attested, and might be reintroduced later.

Further, the Ishtar Gate is Neo-Babylonian and predates the 0 A.D.
timeline. As such, given its limited use, it is removed as a buildable
structure.

Patch by: Nescio
Accepted by: borg-, wraitii
Differential Revision: https://code.wildfiregames.com/D2742
This was SVN commit r23888.