Move gaia/fauna_hawk -> birds/buzzard.
The "hawk" is actually a buzzard, so renamed.
Moved outside of the "gaia" folder so mapmakers can easily identify
between birds that cannot be interacted with ("birds" folder) and those
that *can* be interacted with ("gaia" folder) even though they may have
the same template name.

Patch by: @Nescio.
Differential Revision: D2254
This was SVN commit r23977.